feature: 채팅방 기본 셋팅
