+++ src/main/java/com/growith/tailo/chat/room/ChatRoomRepository.java
@@ -1,0 +1,6 @@
+package com.growith.tailo.chat.room;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
+}
