--- src/main/java/com/growith/tailo/chat/room/controller/ChatRoomController.java
@@ -1,16 +1,0 @@
-package com.growith.tailo.chat.room.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping
-@RequiredArgsConstructor
-public class ChatRoomController {
-    // 채팅방 생성
-
-    //채팅방 조회
-
-    // 채팅방 삭제
-}
