--- src/main/java/com/growith/tailo/chat/room/repository/ChatRoomRepository.java
@@ -1,9 +1,0 @@
-package com.growith.tailo.chat.room.repository;
-
-import com.growith.tailo.chat.room.entity.ChatRoom;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
-}
