--- src/main/java/com/growith/tailo/chat/room/service/ChatRoomService.java
@@ -1,17 +1,0 @@
-package com.growith.tailo.chat.room.service;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class ChatRoomService {
-    //채팅방 생성
-
-    //채팅방 조회
-
-    //채팅방 삭제
-
-}
