--- src/main/java/com/growith/tailo/chat/roommemeber/ChatRoomMemberService.java
@@ -1,13 +1,0 @@
-package com.growith.tailo.chat.roommemeber;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class ChatRoomMemberService {
-    // 조회 ?
-
-}
